+++ proglearn/__init__.py
@@ -1,5 +1,4 @@
 from .forest import *
 from .network import *
-from .oblique_tree import *
 
 __version__ = "0.0.3"
